dss/pyproject.toml
[project]
name = "dss"
version = "0.1.0"
description = "Command line tool for data archives"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.0.0",
    "colorama>=0.4.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
dss = "dss.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "ISC", "PIE", "T20", "Q"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=dss --cov-report=term-missing"
